praca nad CarsDataService - projekt odpala

+++ src/main/java/pl/sda/model/BaseModel.java
@@ -8,16 +8,14 @@ public class BaseModel implements Serializable{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
-
-
 }
